+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- GradientDescentTrainer makes `serialization_dir` when it's instantiated, if it doesn't exist.
+
 ### Added
 
 - Added `transformer_kwargs` argument to `PretrainedTransformerBackbone`
@@ -96,7 +98,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Torch version bumped to 1.7.1 in Docker images.
 - 'master' branch renamed to 'main'
 - Torch version bumped to 1.7.1 in Docker images.
-- GradientDescentTrainer makes `serialization_dir` when it's instantiated, if it doesn't exist.
 
 ### Removed