+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
 - GradientDescentTrainer makes `serialization_dir` when it's instantiated, if it doesn't exist.
 
 ### Added